feat(topology/instances/real): add continuous_at.div
When proving a particular function continuous at a particular point,
lemmas such as continuous_at.mul, continuous_at.add and
continuous_at.comp can be used to build this up from continuity
properties of simpler functions.  It's convenient to have something
similar for division as well.

Note that this only adds continuous_at.div for ℝ, as that was the
context in which I needed it.  The same result is true for other types
such as ℂ and nnreal; if mathlib gets topological (semi)fields in
future, this should become a result for those.