fix(ci): check nolints.txt against master (#2906)
Currently, leanprover-community-bot makes an "update nolints" PR if both of the following hold:
- the `nolints` branch doesn't exist, meaning there isn't already an open nolints PR
- `nolints.txt` has been modified compared to HEAD.

This can lead to a duplicate nolints PR (see e.g. #2899 and #2901), since a nolints PR might have been merged after this build on `master` started, but before this step runs.

This PR changes the second check so that it instead compares `nolints.txt` against the most recent `master` commit, which should fix this.