fix(tactic/squeeze_simp): make `squeeze_simp [←...]` work
`squeeze_simp` parses the argument list using a function in core Lean, and when
support for backwards arguments was added to `simp`, it used a new function to
parse the additional structure. This PR fixes the TODO left in the code to
switch `squeeze_simp` to the new function too, allowing use of backwards
arguments.